+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Bundle the extension
 - Grey out expand/collapse buttons instead of hiding them
+- Don't export status node in tree
 
 ## v0.0.179 - 2020-07-22